Adds PendingUrlRequests and PendingUrlRequest.
Both classes will be used by OlpClient to merge same URL requests and
thus avoid requesting the same data multiple times. For this a change
is needed in HttpResponse to handle the std::stringstream copy on
copy constructor and copy assignment operator. Without this it is not
possible to copy an existing HttpResponse and return it to multiple
waiting callbacks. At the same time we deprecate the public members
of HttpResponse and prepare it to be changed by 01.2020 into a shared
pimpl class so that we do not need to deep copy the responses.
The PendingUrlRequest class manages one Network request, for which it
remebers the request id and is able to cancel it, and is able to hold
one or multiple response callback. It also handles correctly one, more
or all callbacks cancellation scenarios, incl. cancelling the network
request once all pending callbacks are cancelled.

Both PendingUrlRequests and PendingUrlRequest are mutex protected and
thread safe, additionaly PendingUrlRequests class takes care to not
get a conflict in case a URL request was cancelled fully and there is
a new request comming for the same URL. In this case a new request will
be triggered and the old one will be taken out from the pending request
and placed in a separate container for cacelled requests only.

Relates-To: OLPEDGE-1805

Signed-off-by: Andrei Popescu <andrei.popescu@here.com>